+++ src/backend/core/tests/fixtures/matrix.py
@@ -3,6 +3,54 @@
 from rest_framework import status
 
 
+# SEARCH
+def mock_search_empty():
+    """Mock response when no Matrix user has been found through search."""
+    return {
+        "message": {"limited": "false", "results": []},
+        "status_code": status.HTTP_200_OK,
+    }
+
+
+def mock_search_successful(user):
+    """Mock response when exactly one user has been found through search."""
+    return {
+        "message": {
+            "limited": "false",
+            "results": [
+                {
+                    "user_id": f"@{user.email.replace('@', '-')}:user_server.com",
+                    "display_name": f"@{user.name} [Fake]",
+                    "avatar_url": "null",
+                },
+            ],
+        },
+        "status_code": status.HTTP_200_OK,
+    }
+
+
+def mock_search_successful_multiple(user):
+    """Mock response when more than one user has been found through search."""
+    return {
+        "message": {
+            "limited": "false",
+            "results": [
+                {
+                    "user_id": f"@{user.email.replace('@', '-')}:user_server1.com",
+                    "display_name": f"@{user.name} [Fake]",
+                    "avatar_url": "null",
+                },
+                {
+                    "user_id": f"@{user.email.replace('@', '-')}:user_server2.com",
+                    "display_name": f"@{user.name} [Other Fake]",
+                    "avatar_url": "null",
+                },
+            ],
+        },
+        "status_code": status.HTTP_200_OK,
+    }
+
+
 # JOIN ROOMS
 def mock_join_room_successful(room_id):
     """Mock response when succesfully joining room. Same response if already in room."""
@@ -39,7 +87,7 @@ def mock_invite_user_already_in_room(user):
     return {
         "message": {
             "errcode": "M_FORBIDDEN",
-            "error": f"{user.email.replace('@', ':')} is already in the room.",
+            "error": f"{user.email.replace('@', '-')}:home_server.fr is already in the room.",
         },
         "status_code": status.HTTP_403_FORBIDDEN,
     }
@@ -56,7 +104,7 @@ def mock_kick_user_forbidden(user):
     return {
         "message": {
             "errcode": "M_FORBIDDEN",
-            "error": f"You cannot kick user @{user.email.replace('@', ':')}.",
+            "error": f"You cannot kick user @{user.email.replace('@', '-')}.",
         },
         "status_code": status.HTTP_403_FORBIDDEN,
     }